// src/api/types/OpenAiModelReasoningEffort.ts
// This file was auto-generated by Fern from our API Definition.

/**
 * Constrains effort on reasoning for reasoning models (o1, o3, etc.).
 * Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response.
 *
 * @default "medium"
 * @see https://platform.openai.com/docs/api-reference/chat/create#chat-create-reasoning_effort
 */
export const OpenAiModelReasoningEffort = {
    Low: "low",
    Medium: "medium",
    High: "high",
} as const;
export type OpenAiModelReasoningEffort = (typeof OpenAiModelReasoningEffort)[keyof typeof OpenAiModelReasoningEffort];